news_bot/requirements.txt
flask
python-dotenv
crewai
langchain-google-genai
